feat: carregando a foto de perfil também no header
