+++ src/components/Header.tsx
@@ -24,7 +24,10 @@ export default function Header() {
               </p>
             </div>
 
-            <div className='w-10 h-10 rounded-full bg-gray-300 shrink-0'>
+            <div className='w-10 h-10 rounded-full shrink-0 overflow-hidden'>
+                {usuario.fotoPerfil && (
+                    <img src={usuario.fotoPerfil} alt={usuario.nome} className='w-full h-full object-cover' />
+                )}
             </div>
         </header>
     );
